Significant update to the documentation, adding detail about usage, recipe specifications, tool specifications.
Also added a fix to specify if a cookbook is trusted when added. More testing needed.
Changed the file check format for tools so file checks are no longer lists of files, but one file per check - which also removed the need for naming of a file check and simplified the format of the file.