Adds the msl build --install option.
The --install option allows a user to build & install directly to a specific directory. This is a better experience than having to find and copy the install materials from ~/.mussels/install/{target} but may also be a necessity for projects that are not easily relocatable after install, even when using `otool` or `patchelf`.

Bumped version number to 0.2.0, because this is a breaking change for recipes that used `{install}/{target}` to reference the install directory. All of these recipes will have to remove the "`/{target}`".