Fix #20, overzealous tool version pruning
The _get_recipe_version() method was supposed to select a recipe
version, and then prune the list of available tools based on any
restrictions listed in the recipe.  Instead, it was pruning the list of
available tools based on any restrictions listed in ALL recipes, which
naturally caused problems once recipes were added that have
non-overlapping tool version requiirements.

This patch fixes this pruning behavior and also adds much needed debug
log output when pruning occurs and much easier to read error output in
the event that suitable tool versions cannot be found.