Fix check to ensure install files are installed
The glob.glob() function is used to evaluate `*` globs in install file
paths.  It does this by returning a list of actual files found on disk,
meaning that our check to see if the file exists does nothing.  As a
result, missing install files won't fail the build.

This fix makes sure that at least one file is installed.