Fix #25, utf-8 decoding crash
A user reported that Mussels seems to crash when trying to decode output
from a script. While I'm not convinced that this fix is technically
correct so as to properly transcode & render output from the locale of
the machine in question, the fix should make the build complete.

I would definitely appreciate a more correct solution if anyone has one.