Add tool variables for use in recipe scripts
There is no way to make a recipe which has strings in it that vary
depending on which version of a tool is being used.

This commit adds tool variables, which is an optional list of variables
you place under each "platform" in a tool YAML file.

This feature was requested for the clamav project so that recipes which
use cmake and visualstudio can support both visualstudio 2017+ while
specifying the CMake generator string for the specific version of Visual
Studio.

Eg: "-G Visual Studio 15 2017" can now be evaluated from:
    "-G {visualstudio.cmake_generator}"