Add a basic test for tool-variables
Adds a basic test to validate that tool variables propagate to recipes
that use the tools. This is the first Mussels test for doing a build,
and may be a good model for future tests.

In addition fixed a log level issue. The default log level is DEBUG for
all mussels things (running mussels, building recipes, and finding
tools). If you change it to INFO, it presently only changes the level
for the main mussels logger, not for the recipe or tool objects. This
commit has a fix to propagate the log level setting down to the recipes
and tools.

Also, enable tests for pull requests

Add print debugging to try to understand test failure